requirements.txt
# Dependencies for the AI Agent Challenge project
pandas
pdfplumber
pytest
tqdm
requests
python-dotenv
google-generativeai
langgraph